debug, reste LLONG_MIN, et inchallah les floats

+++ src/conversions.c
@@ -6,7 +6,7 @@
 /*   By: cmiran <cmiran@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/30 22:51:31 by cmiran            #+#    #+#             */
-/*   Updated: 2019/04/23 14:13:45 by cmiran           ###   ########.fr       */
+/*   Updated: 2019/04/25 17:24:31 by cmiran           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,9 +32,9 @@ void	is_ouxp(size_t *fla, unsigned long long *unbr, va_list ap, char b)
 	else if (fla['h'])
 		*unbr = (unsigned short)va_arg(ap, unsigned int);
 	else if (fla['l'])
-		*unbr = va_arg(ap, unsigned long);
+		*unbr = (uintmax_t)va_arg(ap, unsigned long);
 	else if (fla[93])
-		*unbr = va_arg(ap, unsigned long long);
+		*unbr = (uintmax_t)va_arg(ap, unsigned long long);
 	else
 		*unbr = va_arg(ap, unsigned int);
 }
